# src/test/resources/openapi/collectra-api-v1-baseline.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Collectra API",
    "description" : "Tenant-aware collection and communications platform",
    "version" : "v1"
  },
  "paths" : {
...
    "/api/v1/template-builder/versions/{versionId}" : {
      "get" : {
        "tags" : [ "template-builder-controller" ],
        "operationId" : "getVersion",
        "parameters" : [ {
          "name" : "versionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "template-builder-controller" ],
        "operationId" : "updateDraft",
        "parameters" : [ {
          "name" : "versionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DraftRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collection-cases/{caseId}" : {
      "get" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "get_2",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "update_5",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaseUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collection-cases" : {
      "get" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "list_10",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "invoiceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "OPEN", "IN_PROGRESS", "ON_HOLD", "CLOSED" ]
          }
        }, {
          "name" : "priority",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "LOW", "NORMAL", "HIGH", "URGENT" ]
          }
        }, {
          "name" : "assignedTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50,
            "maximum" : 200,
            "minimum" : 1
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt,desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CasePage"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "create_11",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaseCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/start" : {
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "start",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/promises" : {
      "get" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "promises",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PromiseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "createPromise",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PromiseCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromiseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/promises/{promiseId}/fulfill" : {
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "fulfillPromise",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "promiseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromiseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/promises/{promiseId}/cancel" : {
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "cancelPromise",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "promiseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromiseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/promises/{promiseId}/break" : {
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "breakPromise",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "promiseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PromiseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/hold" : {
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "hold",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collection-cases/{caseId}/disputes" : {
      "get" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "disputes",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DisputeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "collection-controller" ],
        "operationId" : "createDispute",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DisputeCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DisputeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "VersionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "templateId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "locale" : {
            "type" : "string"
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "EMAIL", "SMS", "WHATSAPP", "TELEGRAM", "PDF" ]
          },
          "subject" : {
            "type" : "string"
          },
          "contentHtml" : {
            "type" : "string"
          },
          "stylesheet" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "VALIDATED", "PUBLISHED", "ARCHIVED" ]
          }
        }
      },
...
      "DraftRequest" : {
        "type" : "object",
        "properties" : {
          "channel" : {
            "type" : "string",
            "enum" : [ "EMAIL", "SMS", "WHATSAPP", "TELEGRAM", "PDF" ]
          },
          "locale" : {
            "type" : "string",
            "maxLength" : 10,
            "minLength" : 0
          },
          "subject" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "content" : {
            "type" : "string",
            "minLength" : 1
          },
          "stylesheet" : {
            "type" : "string"
          }
        },
        "required" : [ "content", "locale" ]
      },
...
      "CaseUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "version" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "LOW", "NORMAL", "HIGH", "URGENT" ]
          },
          "assignedTo" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
    }
  }
}